moved some funcs
